Expression Tree functionality is now working.
Had some weird issues getting the templates to play nice... template specialization is kind of weird since there's no way to do partial template specialization which means you have to provide the template type in every case when its instantiated.

Also had a typo in the construction of the Expression Tree that was giving me a bunch of null pointer exception bugs, caused by the fact that I forgot to instantiate a right node for a root node.

Code still needs lots of cleanup and removing old comments, and the BST->LL functionality needs to be tested.

Signed-off-by: Raleigh Littles III <raleighlittles@gmail.com>