First draft works mostly.
The issue was because the RapidCSV parser expects there to be a row header as well as a column header, so it was ignoring the first column (that had the words in it).

Since my dictionary.csv file has words starting with capital letters, I should probably change the lookup() method to capitalize the word so that users don't have to remember to do that.

Signed-off-by: Raleigh Littles III <raleighlittles@gmail.com>