First commit for Chapter 13 Problem #12.
Added a simple function that takes in a list of words and returns the length of the longest word in that vector.
Been reviewing different approaches to use, think I figured out a way to convert the regular radix sort for integers to be compatible with strings by simply using the ASCII value of a character as its sorting characteristic and then making buckets for every possible ASCII character.

Signed-off-by: Raleigh Littles III <raleighlittles@gmail.com>